# app/microservice/public-swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GLobal Forest Watch API",
    "description": "The Global Forest Watch API is currently in beta. Based on feedback from trusted testers, this document specifies new API changes that are currently under development. Interested in being a trusted tester? Contact the [Data Lab](mailTo:support@wri.org) at World Resources Institute.\n\nAll endpoints are accessible via http & https and are located at\n`https://production-api.globalforestwatch.org`.\n\nForest change measures tree cover loss, tree cover gain, or forest disturbance. The following forest change data are available through the API: \n* University of Maryland tree cover loss & gain \n* FORMA alerts\n* IMAZON SAD alerts\n* Terrai alerts\n* Prodes\n* Guira Loss\n* Biomass loss\n* GLAD tree cover loss alerts \n* NASA VIIRs active fires.\n\nOther services provided by the Api complements the ecosystem that surrounds the platform:\n\n* Metadata\n* Urthecast \n* Geostore\n* OGR \n* Story \n* User \n* subscription\n\n\n### Forest Change - Viirs Fires Alerts API\nDisplays fire alert data for the past 7 days.\n\n| Property | Value |\n| --- | --- | --- |\n| Coverage | Global |\n| Resolution | 1 x 1 kilometer |\n| Source | MODIS |\n| Timescale | Last 7 days |\n| Units | Alerts |\n| Updates | Daily |\n|Source| http://viirsfire.geog.umd.edu/|\n",
    "termsOfService": "http://www.globalforestwatch.org/terms",
    "contact": {
      "name": "API Support",
...
      "email": "support@wri.org"
...
  "host": "staging-api.globalforestwatch.org",
  "schemes": [
...
  "consumes": [
...
  "produces": [
...
  "basePath": "/v1",
  "paths": {
    "/viirs-active-fires": {
      "get": {
        "description": "Retrives total number of alerts for your custom geometry on the period selected\n",
        "operationId": "world",
        "consumes": [
          "application/vnd.api+json"
        ],
        "produces": [
          "application/vnd.api+json",
          "application/json"
        ],
        "tags": [
          "VIIRS-FIRES"
        ],
        "parameters": [
          {
            "name": "geostore",
            "in": "query",
            "description": "Geostore hash \n```34cecd4c4809dbbec20ce40be3d1c877```\n",
            "required": true,
            "type": "string"
          },
          {
            "name": "period",
            "in": "query",
            "description": "Period of time. NB: This MS queries CARTO table vnp14imgtdl_nrt_global_7d, which is limited to the last 7 days. This period filter will applied when used, but it can't query data that doesn't exist. As such, any queries outside of the last 7 days will return zero fires. \n```2016-12-01,2016-12-07```\n",
            "required": false,
            "default": "past 24 hours",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Results obteined",
            "examples": {
              "data": {
                "type": "viirs-alerts",
                "id": "f123fswse13",
                "attributes": {
                  "value": 988,
                  "areaHa": 10472.639768453402,
                  "downloadUrls": {
                    "csv": "/download/274b4818-be18-4890-9d10-eae56d2a82e5?sql=select lat, lon, confidence, year, julian_day from index_b846230fcec0420892d13fc11ea7e32b where year = 2016 and julian_day >= 1 and julian_day <= 347&geostore=34cecd4c4809dbbec20ce40be3d1c877&format=csv",
                    "json": "/download/274b4818-be18-4890-9d10-eae56d2a82e5?sql=select lat, lon, confidence, year, julian_day from index_b846230fcec0420892d13fc11ea7e32b where year = 2016 and julian_day >= 1 and julian_day <= 347&geostore=34cecd4c4809dbbec20ce40be3d1c877&format=json"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/definitions/VIIRSFIRESAPI"
            }
          },
          "404": {
            "description": "Geostore not found",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          }
        }
      },
      "post": {
        "description": "Retrives total number of alerts for your custom geometry on the period selected\n",
        "operationId": "worldWithGeojson",
        "consumes": [
          "application/vnd.api+json"
        ],
        "produces": [
          "application/vnd.api+json",
          "application/json"
        ],
        "tags": [
          "VIIRS-FIRES"
        ],
        "parameters": [
          {
            "name": "geojson",
            "in": "body",
            "description": "##### examples\n  ```json\n  {\n      \"geojson\": {\n          \"type\": \"FeatureCollection\",\n          \"features\": [{\n              \"type\": \"Feature\",\n              \"properties\": {},\n              \"geometry\": {\n                  \"type\": \"Polygon\",\n                  \"coordinates\": [\n                      [\n                          [-5.273512601852417, 42.81137220349083],\n                          [-5.273512601852417, 42.811803118457306],\n                          [-5.272732079029083, 42.811803118457306],\n                          [-5.272732079029083, 42.81137220349083],\n                          [-5.273512601852417, 42.81137220349083]\n                      ]\n                  ]\n              }\n          }]\n      }\n  }\n  ```\n",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GeoJsonBodyRequest"
            }
          },
          {
            "name": "period",
            "description": "Period of time. NB: This MS queries CARTO table vnp14imgtdl_nrt_global_7d, which is limited to the last 7 days. This period filter will applied when used, but it can't query data that doesn't exist. As such, any queries outside of the last 7 days will return zero fires. \n```2016-12-01,2016-12-07```\n",
            "required": false,
            "default": "past 24 hours",
            "type": "string",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Results obteined",
            "examples": {
              "data": {
                "type": "viirs-alerts",
                "id": "f123fswse13",
                "attributes": {
                  "value": 988,
                  "areaHa": 10472.639768453402,
                  "downloadUrls": {
                    "csv": "/download/274b4818-be18-4890-9d10-eae56d2a82e5?sql=select lat, lon, confidence, year, julian_day from index_b846230fcec0420892d13fc11ea7e32b where year = 2016 and julian_day >= 1 and julian_day <= 347&geostore=34cecd4c4809dbbec20ce40be3d1c877&format=csv",
                    "json": "/download/274b4818-be18-4890-9d10-eae56d2a82e5?sql=select lat, lon, confidence, year, julian_day from index_b846230fcec0420892d13fc11ea7e32b where year = 2016 and julian_day >= 1 and julian_day <= 347&geostore=34cecd4c4809dbbec20ce40be3d1c877&format=json"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/definitions/VIIRSFIRESAPI"
            }
          },
          "404": {
            "description": "Geostore not found",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          }
        }
      }
    },
    "/viirs-active-fires/use/{name}/{id}": {
      "get": {
        "operationId": "use",
        "description": "Retrives total number of alerts for your selected concession geometry on the period selected\n",
        "produces": [
          "application/vnd.api+json"
        ],
        "tags": [
          "VIIRS-FIRES"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Concession type, one of  \n```mining, oilpalm, fiber, logging```\n",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Concession geometry id\n``` 100 ```\n",
            "required": true,
            "type": "integer"
          },
          {
            "name": "period",
            "in": "query",
            "description": "Period of time. NB: This MS queries CARTO table vnp14imgtdl_nrt_global_7d, which is limited to the last 7 days. This period filter will applied when used, but it can't query data that doesn't exist. As such, any queries outside of the last 7 days will return zero fires. \n```2016-12-01,2016-12-07```\n",
            "required": false,
            "default": "past 24 hours",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Results obtained",
            "examples": {
              "data": {
                "type": "viirs-fires",
                "id": "f123fswse13",
                "attributes": {
                  "value": 988,
                  "areaHa": 10472.639768453402,
                  "downloadUrls": {
                    "csv": "/download/274b4818-be18-4890-9d10-eae56d2a82e5?sql=select lat, lon, confidence, year, julian_day from index_b846230fcec0420892d13fc11ea7e32b where year = 2016 and julian_day >= 1 and julian_day <= 347&geostore=34cecd4c4809dbbec20ce40be3d1c877&format=csv",
                    "json": "/download/274b4818-be18-4890-9d10-eae56d2a82e5?sql=select lat, lon, confidence, year, julian_day from index_b846230fcec0420892d13fc11ea7e32b where year = 2016 and julian_day >= 1 and julian_day <= 347&geostore=34cecd4c4809dbbec20ce40be3d1c877&format=json"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/definitions/VIIRSFIRESAPI"
            }
          },
          "404": {
            "description": "Geostore not found",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          }
        }
      }
    },
    "/viirs-active-fires/wdpa/{id}": {
      "get": {
        "operationId": "wdpa",
        "description": "Retrives total number of alerts for your selected protected area on the period selected\n",
        "produces": [
          "application/vnd.api+json"
        ],
        "tags": [
          "VIIRS-FIRES"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "wdpa id\n``` 56670 ```\n",
            "required": true,
            "type": "integer"
          },
          {
            "name": "period",
            "in": "query",
            "description": "Period of time. NB: This MS queries CARTO table vnp14imgtdl_nrt_global_7d, which is limited to the last 7 days. This period filter will applied when used, but it can't query data that doesn't exist. As such, any queries outside of the last 7 days will return zero fires. \n```2016-12-01,2016-12-07```\n",
            "required": false,
            "default": "past 24 hours",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Results obtained",
            "examples": {
              "data": {
                "type": "viirs-fires",
                "id": "f123fswse13",
                "attributes": {
                  "value": 988,
                  "areaHa": 10472.639768453402,
                  "downloadUrls": {
                    "csv": "/download/274b4818-be18-4890-9d10-eae56d2a82e5?sql=select lat, lon, confidence, year, julian_day from index_b846230fcec0420892d13fc11ea7e32b where year = 2016 and julian_day >= 1 and julian_day <= 347&geostore=34cecd4c4809dbbec20ce40be3d1c877&format=csv",
                    "json": "/download/274b4818-be18-4890-9d10-eae56d2a82e5?sql=select lat, lon, confidence, year, julian_day from index_b846230fcec0420892d13fc11ea7e32b where year = 2016 and julian_day >= 1 and julian_day <= 347&geostore=34cecd4c4809dbbec20ce40be3d1c877&format=json"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/definitions/VIIRSFIRESAPI"
            }
          },
          "404": {
            "description": "Geostore not found",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          }
        }
      }
    },
    "/viirs-active-fires/admin/{iso}": {
      "get": {
        "operationId": "getNational",
        "description": "Retrives total number of alerts for your selected country on the period selected\n",
        "produces": [
          "application/vnd.api+json"
        ],
        "tags": [
          "VIIRS-FIRES"
        ],
        "parameters": [
          {
            "name": "iso",
            "in": "path",
            "description": "ISO-3 country code \n``` BRA ```\n",
            "required": true,
            "type": "string"
          },
          {
            "name": "period",
            "in": "query",
            "description": "Period of time. NB: This MS queries CARTO table vnp14imgtdl_nrt_global_7d, which is limited to the last 7 days. This period filter will applied when used, but it can't query data that doesn't exist. As such, any queries outside of the last 7 days will return zero fires. \n```2016-12-01,2016-12-07```\n",
            "required": false,
            "default": "past 24 hours",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Results obtained",
            "examples": {
              "data": {
                "type": "viirs-fires",
                "id": "f123fswse13",
                "attributes": {
                  "value": 988,
                  "areaHa": 10472.639768453402,
                  "downloadUrls": {
                    "csv": "/download/274b4818-be18-4890-9d10-eae56d2a82e5?sql=select lat, lon, confidence, year, julian_day from index_b846230fcec0420892d13fc11ea7e32b where year = 2016 and julian_day >= 1 and julian_day <= 347&geostore=34cecd4c4809dbbec20ce40be3d1c877&format=csv",
                    "json": "/download/274b4818-be18-4890-9d10-eae56d2a82e5?sql=select lat, lon, confidence, year, julian_day from index_b846230fcec0420892d13fc11ea7e32b where year = 2016 and julian_day >= 1 and julian_day <= 347&geostore=34cecd4c4809dbbec20ce40be3d1c877&format=json"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/definitions/VIIRSFIRESAPI"
            }
          },
          "404": {
            "description": "Geostore not found",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          }
        }
      }
    },
    "/viirs-active-fires/admin/{iso}/{id1}": {
      "get": {
        "operationId": "getSubnational",
        "description": "Retrives total number of alerts for your selected admin 1 area based on GADM 2 on the period selected\n",
        "produces": [
          "application/vnd.api+json"
        ],
        "tags": [
          "VIIRS-FIRES"
        ],
        "parameters": [
          {
            "name": "iso",
            "in": "path",
            "description": "ISO-3 country code \n``` BRA ```\n",
            "required": true,
            "type": "string"
          },
          {
            "name": "id1",
            "in": "path",
            "description": "numeric code for admin 1 region code based on GADM 2 \n``` 2 ```\n",
            "required": true,
            "type": "integer"
          },
          {
            "name": "period",
            "in": "query",
            "description": "Period of time. NB: This MS queries CARTO table vnp14imgtdl_nrt_global_7d, which is limited to the last 7 days. This period filter will applied when used, but it can't query data that doesn't exist. As such, any queries outside of the last 7 days will return zero fires. \n```2016-12-01,2016-12-07```\n",
            "required": false,
            "default": "past 24 hours",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Results obtained",
            "schema": {
              "$ref": "#/definitions/VIIRSFIRESAPI"
            }
          },
          "404": {
            "description": "Geostore not found",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          }
        }
      }
    },
    "/viirs-active-fires/latest": {
      "get": {
        "operationId": "getLatest",
        "description": "Retrives latest alerts\n",
        "produces": [
          "application/vnd.api+json"
        ],
        "tags": [
          "VIIRS-FIRES"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Limit. Default 3",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Results obtained",
            "schema": {
              "$ref": "#/definitions/VIIRSLATESTAPI"
            }
          },
          "404": {
            "description": "Geostore not found",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/Errors"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Errors": {
      "type": "object",
      "description": "Errors",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "A unique identifier for this particular occurrence of the problem."
        },
        "links": {
          "type": "object",
          "description": "A links object",
          "properties": {
            "about": {
              "type": "string",
              "description": "A link that leads to further details about this particular occurrence of the problem."
            }
          }
        },
        "status": {
          "type": "string",
          "description": "The HTTP status code applicable to this problem, expressed as a string value"
        },
        "code": {
          "type": "string",
          "description": "An application-specific error code, expressed as a string value"
        },
        "title": {
          "type": "string",
          "description": "A short, human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the problem, except for purposes of localization."
        },
        "detail": {
          "type": "string",
          "description": "A human-readable explanation specific to this occurrence of the problem. Like title, this field's value can be localized"
        },
        "source": {
          "type": "object",
          "description": "An object containing references to the source of the error, optionally including any of the following members",
          "properties": {
            "pointer": {
              "type": "string",
              "description": "A JSON Pointer [RFC6901] to the associated entity in the request document"
            },
            "parameter": {
              "type": "string",
              "description": "A string indicating which URI query parameter caused the error."
            }
          }
        },
        "meta": {
          "type": "object",
          "description": "A meta object containing non-standard meta-information about the error."
        }
      }
    },
    "VIIRSLATESTAPI": {
      "type": "object",
      "properties": {
        "data": {
          "properties": {
            "schema": {
              "$ref": "#/definitions/VIIRSLATESTBody"
            }
          }
        }
      }
    },
    "VIIRSLATESTBody": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "description": "Resource type"
        },
        "id": {
          "type": "string",
          "description": "Resource id"
        },
        "attributes": {
          "properties": {
            "schema": {
              "$ref": "#/definitions/VIIRSLATESTModel"
            }
          }
        }
      }
    },
    "VIIRSLATESTModel": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "format": "date"
        }
      }
    },
    "VIIRSFIRESAPI": {
      "type": "object",
      "properties": {
        "data": {
          "properties": {
            "schema": {
              "$ref": "#/definitions/VIIRSFIRESBody"
            }
          }
        }
      }
    },
    "VIIRSFIRESBody": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "description": "Resource type"
        },
        "id": {
          "type": "string",
          "description": "Resource id"
        },
        "attributes": {
          "properties": {
            "schema": {
              "$ref": "#/definitions/VIIRSFIRESModel"
            }
          }
        }
      }
    },
    "VIIRSFIRESModel": {
      "type": "object",
      "properties": {
        "value": {
          "type": "number",
          "format": "float"
        },
        "period": {
          "type": "string"
        },
        "download-urls": {
          "type": "object",
          "properties": {
            "csv": {
              "type": "string"
            },
            "geojson": {
              "type": "string"
            },
            "kml": {
              "type": "string"
            },
            "shp": {
              "type": "string"
            },
            "svg": {
              "type": "string"
            }
          }
        }
      }
    },
    "GeoJsonBodyRequest": {
      "type": "object",
      "properties": {
        "geojson": {
          "properties": {
            "schema": {
              "$ref": "#/definitions/Features"
            }
          }
        }
      }
    },
    "Features": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "description": "Type of feature"
        },
        "features": {
          "type": "array",
          "description": "Features",
          "items": {
            "$ref": "#/definitions/Feature"
          }
        }
      }
    },
    "Feature": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "description": "Type of feature"
        },
        "properties": {
          "type": "object",
          "description": "Feature"
        },
        "geometry": {
          "type": "object",
          "description": "Geometries",
          "properties": {
            "schema": {
              "$ref": "#/definitions/Geometry"
            }
          }
        }
      }
    },
    "Geometry": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "description": "Geometry type"
        },
        "coordinates": {
          "type": "array",
          "description": "Coordinates array",
          "items": {
            "type": "object"
          }
        }
      }
    }
  }
}